net: dsa: mv88e6xxx: add PVT support
Some Marvell switch chips (such as 6352 and 6390) have a Cross-chip Port
Based VLAN Table (PVT), which is accessed by the DSA links to ensure
that frames from another switch port are allowed to egress on a given
local port.

The default (and current) configuration is to set it to all ones,
allowing any external ports to egress frames on local ports.

This commit implements the PVT access functions and assigned them to
chips supporting it. The current behavior is unchanged for the moment.

Signed-off-by: Vivien Didelot <vivien.didelot@savoirfairelinux.com>