net: dsa: mv88e6xxx: rework in-chip bridging
Now that DSA exposes the number of ports of other switches and both
bridge join and leave operations pass the bridge device, it is possible
to factorize the port VLAN mapping code a bit.

Also, make the mv88e6xxx_port_vlan() helper handle the logic of any
switch port, since this will be used by the cross-chip bridging code.

Signed-off-by: Vivien Didelot <vivien.didelot@savoirfairelinux.com>