net: dsa: rearrange slave EEE calls
First the PHY EEE settings need to be accessed, then the MAC EEE
settings. Then, a DSA driver may not implement the port's MAC EEE access
routines but still expose PHY devices. But if there is no PHY device
related to the slave port, we must not call driver operations.

This patch rearrange the slave get and set EEE routines to respect that.

Signed-off-by: Vivien Didelot <vivien.didelot@savoirfairelinux.com>