net: dsa: move master ethtool ops in dsa_master
The copy of the master netdev ethtool_ops is currently stored in the
dsa_port structure, but it is specific to the DSA master (CPU) port.

Move it in the new dsa_master structure. While modifying the ethtool
functions to take this structure as argument, move them in a new
master.c DSA core file.

Also the ethtool ops were the only remaining users for
master->port->netdev thus remove this assignment now.

Signed-off-by: Vivien Didelot <vivien.didelot@savoirfairelinux.com>