net: dsa: change dsa_ptr for a dsa_master
The dsa_ptr of a net_device currently points to a dsa_switch_tree.

This is not correct because the interface is physically linked to a
switch port, and because several network interface could point to
several (so called "CPU") ports of the same DSA switch tree.

Because the tagging operations are specific to the DSA master port and
not the whole switch tree, move the tag_ops pointer in dsa_master.

Finally change the dsa_ptr from dsa_switch_tree to a dsa_master.

Signed-off-by: Vivien Didelot <vivien.didelot@savoirfairelinux.com>