net: dsa: add dsa_is_unused_port helper
As the comment above the chunk states, the drivers attempts to disable
the unused port. But using ds->enabled_port_mask is misleading, because
this mask reports in fact the user ports.

To avoid confusion and fix this, this patch introduces an explicit
dsa_is_unused_port helper which ensures the corresponding bit is not
masked in any of the switch port masks.

Signed-off-by: Vivien Didelot <vivien.didelot@savoirfairelinux.com>