net: dsa: add find port by node helper
Instead of having two dsa_ds_find_port_dn (which returns a bool) and
dsa_dst_find_port_dn (which returns a switch) functions, provide a more
explicit dsa_tree_find_port_by_node function which returns a matching
port.

The OF code provides a helper to iterate over phandles. Use it instead
of arbitrary iterating ourselves on the list of phandles hanging to the
"link" property of the port's device node.

The of_parse_phandle() helper calls of_node_put() itself on it.node. We
must only do it ourselves if we break the loop.

Signed-off-by: Vivien Didelot <vivien.didelot@savoirfairelinux.com>