net: dsa: setup and teardown tree
This commit brings no functional changes. It simply provides better
scope for the DSA tree setup and teardown functions.

It also renames the "applied" bool to "setup" and print a message when
the tree is setup, as it is down during teardown.

At the same time, check dst->setup in dsa_tree_setup, where it is set to
true.

Signed-off-by: Vivien Didelot <vivien.didelot@savoirfairelinux.com>