net: dsa: rename netdev_uses_dsa
The netdev_uses_dsa helper is used to check if a given net_device is
used as a DSA master interface. We will soon need to check if a
net_device is a DSA slave as well. So before introducing a
netdev_is_dsa_slave helper, rename netdev_uses_dsa to
netdev_is_dsa_master for consistency.

Signed-off-by: Vivien Didelot <vivien.didelot@savoirfairelinux.com>