net: dsa: mv88e6xxx: add Wait Bit bus operation
Some busses have their own implementation of the Wait Bit
routine. Reflect that in the mv88e6xxx_bus_ops structure, and prefer
it over the generic implementation if one is provided.

The SMI direct wait is already implemented for the indirect variant,
so use it even though it isn't a SMI specific implementation per-se.

Signed-off-by: Vivien Didelot <vivien.didelot@gmail.com>